[googlemaps] Extract geometry.d.ts
